10_context_dynamic_instructions/pyproject.toml
[project]
name = "10-context-dynamic-instructions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
